Refactor: Alteração do arquivo em que a enum CarState é declarada
